refactor processors and provider + updated models
i refactored the providers and the processors to use the SQLite database ILY.db now and updated the models to so that all the endpoint work. this is still not a 100% garente because i have not tested them al because of the fact that i am to tired to test at this moment in time.